added: a separate plugin for handling responsive images
