+++ plugins/content/joomla/joomla.php
@@ -22,7 +22,6 @@ use Joomla\Component\Workflow\Administrator\Table\WorkflowTable;
 use Joomla\Database\DatabaseDriver;
 use Joomla\Database\ParameterType;
 use Joomla\Utilities\ArrayHelper;
-use Joomla\CMS\Helper\MediaHelper;
 
 /**
  * Example Content Plugin
@@ -82,27 +81,6 @@ class PlgContentJoomla extends CMSPlugin
 			return $this->checkMenuItemBeforeSave($context, $table, $isNew, $data);
 		}
 
-		// Check if context not includes com_media - com_media.file or com_media.folder
-		if (explode('.', $context)[0] !== 'com_media')
-		{
-			// Add srcset attribute to content images
-			$contentKey = $this->_getContentKey($context);
-
-			$item = clone $table;
-			$item->load($table->id);
-
-			// Get initial versions of content and form images
-			if ($formImages = $this->_getFormImages($context, (array) $item))
-			{
-				$this->initFormImages = $formImages;
-			}
-
-			if ($content = $item->{$contentKey})
-			{
-				$this->initContent = $content;
-			}
-		}
-
 		// Check we are handling the frontend edit form.
 		if (!in_array($context, ['com_workflow.stage', 'com_workflow.workflow']) || $isNew)
 		{
@@ -142,21 +120,6 @@ class PlgContentJoomla extends CMSPlugin
 	 */
 	public function onContentAfterSave($context, $article, $isNew): void
 	{
-		// Check if context not includes com_media - com_media.file or com_media.folder
-		if (explode('.', $context)[0] !== 'com_media')
-		{
-			// Generate responsive images for form and content
-			if ($formImages = $this->_getFormImages($context, (array) $article))
-			{
-				MediaHelper::generateFormResponsiveImages($this->initFormImages, $formImages);
-			}
-
-			if ($content = $article->{$this->_getContentKey($context)})
-			{
-				MediaHelper::generateContentResponsiveImages($this->initContent, $content);
-			}
-		}
-
 		// Check we are handling the frontend edit form.
 		if ($context !== 'com_content.form')
 		{
@@ -280,61 +243,6 @@ class PlgContentJoomla extends CMSPlugin
 		}
 
 		return true;
-	}
-
-	/**
-	 * Returns form images from data with specific context
-	 *
-	 * @param   string  $context  The context for the data
-	 * @param   array   $data     The validated data
-	 *
-	 * @return  mixed   Array of form images or false if they don't exist
-	 *
-	 * @since   4.1.0
-	 */
-	private function _getFormImages($context, $data)
-	{
-		// Convert string images to array
-		$data['images'] = (array) json_decode($data['images']);
-		$data['params'] = (array) json_decode($data['params']);
-
-		// Get form images depending on context
-		switch ($context)
-		{
-			case "com_content.article":
-			case "com_tags.tag":
-				return array(
-					'image_intro' => $data['images']['image_intro'], 'image_fulltext' => $data['images']['image_fulltext']
-				);
-			case "com_banners.banner":
-				return array('image' => $data['params']['imageurl']);
-			case "com_categories.category":
-				return array('image' => $data['params']['image']);
-			case "com_contact.contact":
-				return array('image' => $data['image']);
-			case "com_newsfeeds.newsfeed":
-				return array(
-					'image_first' => $data['images']['image_first'], 'image_second' => $data['images']['image_second']
-				);
-			default:
-				return false;
-		}
-	}
-
-	/**
-	 * Returns content key in form object for specific context
-	 *
-	 * @param   string  $context  The context
-	 *
-	 * @return  string  Content key
-	 *
-	 * @since   4.1.0
-	 */
-	private function _getContentKey($context)
-	{
-		return $context === 'com_content.article' ? 'introtext' : (
-			$context === 'com_contact.contact' ? 'misc' : 'description'
-		);
 	}
 
 	/**
