+++ plugins/content/responsiveimages/responsiveimages.xml
@@ -1,0 +1,38 @@
+<?xml version="1.0" encoding="utf-8"?>
+<extension type="plugin" group="content" method="upgrade">
+	<name>plg_content_responsiveimages</name>
+	<author>Joomla! Project</author>
+	<creationDate>July 2021</creationDate>
+	<copyright>(C) 2021 Open Source Matters, Inc.</copyright>
+	<license>GNU General Public License version 2 or later; see LICENSE.txt</license>
+	<authorEmail>admin@joomla.org</authorEmail>
+	<authorUrl>www.joomla.org</authorUrl>
+	<version>4.1.0</version>
+	<description>PLG_CONTENT_RESPONSIVEIMAGES_XML_DESCRIPTION</description>
+	<files>
+		<filename plugin="responsiveimages">responsiveimages.php</filename>
+	</files>
+	<languages>
+		<language tag="en-GB">en-GB.plg_content_responsiveimages.ini</language>
+		<language tag="en-GB">en-GB.plg_content_responsiveimages.sys.ini</language>
+	</languages>
+	<config>
+		<fields name="params">
+			<fieldset name="basic">
+				<field
+					name="example"
+					type="radio"
+					layout="joomla.form.field.radio.switcher"
+					label="PLG_CONTENT_RESPONSIVEIMAGES_FIELD_EXAMPLE_LABEL"
+					description="PLG_CONTENT_RESPONSIVEIMAGES_FIELD_EXAMPLE_DESC"
+					default="1"
+					filter="integer"
+					>
+					<option value="0">JNO</option>
+					<option value="1">JYES</option>
+				</field>
+			</fieldset>
+		</fields>
+	</config>
+
+</extension>
